server/requirement.txt
paho-mqtt
pandas
PTL
Pillow
tensorflow==2.17.0
numpy==1.23.5
scikit-learn==1.2.2
nbconvert
nbclient
jupyter_client
pyzmq
ipykernel